refactor: 중복 dto 새로 생성
