+++ src/main/java/org/prgms/locomocoserver/user/dto/response/TokenResponseDto.java
@@ -1,8 +1,10 @@
-package org.prgms.locomocoserver.user.dto.github;
+package org.prgms.locomocoserver.user.dto.response;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public record GithubTokenResponseDto(
+@JsonIgnoreProperties(ignoreUnknown = true)
+public record TokenResponseDto(
         @JsonProperty("access_token") String accessToken,
         @JsonProperty("token_type") String tokenType,
         @JsonProperty("refresh_token") String refreshToken,
