+++ src/main/java/org/prgms/locomocoserver/user/presentation/GithubController.java
@@ -9,8 +9,8 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.prgms.locomocoserver.user.application.UserService;
-import org.prgms.locomocoserver.user.dto.github.GithubTokenResponseDto;
 import org.prgms.locomocoserver.user.dto.github.GithubUserInfoResponseDto;
+import org.prgms.locomocoserver.user.dto.response.TokenResponseDto;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -48,8 +48,8 @@ public class GithubController {
     @ApiResponse(responseCode = "200", description = "로그인 성공")
     @Parameter(name = "code", description = "로그인 정보 입력 시 깃허브에서 반환되는 일회성 code")
     @GetMapping("/users/login/github/callback")
-    public ResponseEntity<GithubTokenResponseDto> getGithubLoginCallback(@RequestParam(name = "code") String code) throws JsonProcessingException {
-        GithubTokenResponseDto githubTokenResponseDto = getTokenDto(code);
+    public ResponseEntity<TokenResponseDto> getGithubLoginCallback(@RequestParam(name = "code") String code) throws JsonProcessingException {
+        TokenResponseDto githubTokenResponseDto = getTokenDto(code);
         GithubUserInfoResponseDto githubUserInfoResponseDto = loadUserInfo(githubTokenResponseDto.accessToken());
         userService.saveOrUpdate(githubUserInfoResponseDto);
 
@@ -64,13 +64,13 @@ public class GithubController {
         return ResponseEntity.ok(userInfo);
     }
 
-    private GithubTokenResponseDto getTokenDto(String code) {
+    private TokenResponseDto getTokenDto(String code) {
         String tokenUrl = "https://github.com/login/oauth/access_token";
 
         RestTemplate restTemplate = new RestTemplate();
-        ResponseEntity<GithubTokenResponseDto> response = restTemplate.postForEntity(
+        ResponseEntity<TokenResponseDto> response = restTemplate.postForEntity(
                 tokenUrl + "?client_id={clientId}&client_secret={clientSecret}&code={code}&redirect_uri={redirectUri}",
-                null, GithubTokenResponseDto.class, github_client_id, github_client_secret_key, code, github_redirect_url);
+                null, TokenResponseDto.class, github_client_id, github_client_secret_key, code, github_redirect_url);
         return response.getBody();
     }
